requirements.txt
asgiref==3.7.2
certifi==2024.12.14
charset-normalizer==3.4.1
crispy-bootstrap5==0.7
dj-database-url==2.2.0
Django==3.2.25
django-crispy-forms==2.0
django-recaptcha==4.0.0
gunicorn==23.0.0
idna==3.10
importlib-metadata==6.7.0
packaging==24.0
psycopg2==2.9.9
python-decouple==3.8
python-dotenv==0.21.1
pytz==2024.2
pywin32==308
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.7
websocket-client==1.6.1
zipp==3.15.0
